+++ app/src/main/java/com/devpaul/ecommerceappmvvm/presentation/MainActivity.kt
@@ -1,0 +1,27 @@
+package com.devpaul.ecommerceappmvvm.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Surface
+import androidx.compose.ui.Modifier
+import com.devpaul.ecommerceappmvvm.presentation.screens.auth.login.LoginScreen
+import com.devpaul.ecommerceappmvvm.presentation.ui.theme.EcommerceAppMVVMTheme
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            EcommerceAppMVVMTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colors.background
+                ) {
+                    LoginScreen()
+                }
+            }
+        }
+    }
+}
