+++ app/src/main/java/com/devpaul/ecommerceappmvvm/presentation/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.devpaul.ecommerceappmvvm.ui.theme
+package com.devpaul.ecommerceappmvvm.presentation.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
