+++ app/src/main/java/com/devpaul/ecommerceappmvvm/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.devpaul.ecommerceappmvvm.ui.theme
+package com.devpaul.ecommerceappmvvm.presentation.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
